refactor(depend): 重构枚举相关代码
- 将 AnyTypingConverterFactory 重命名为 AnyEnumConverterFactory
- 更新相关类和接口名称，从 Typing 改为 Enum
- 修改 HttpServletResponseFns 和相关测试文件，更新 MediaTypes 引用
