+++ depend/servlet/src/main/kotlin/io/github/truenine/composeserver/depend/servlet/HttpServletResponseFns.kt
@@ -1,7 +1,7 @@
 package io.github.truenine.composeserver.depend.servlet
 
 import io.github.truenine.composeserver.consts.IHeaders
-import io.github.truenine.composeserver.typing.MediaTypes
+import io.github.truenine.composeserver.enums.MediaTypes
 import jakarta.servlet.http.HttpServletResponse
 import java.io.OutputStream
 import java.nio.charset.Charset
